requirements.txt
django
django-jazzmin
django-ckeditor
dj-static
pillow
readtime
django-environ==0.10.0
sslcommerz-python
